春

+++ sylzyb_employer_mgr/sylzyb_employer_mgr/employer.aspx.cs
@@ -6,6 +6,8 @@ using System.Web.UI;
 using System.Web.UI.WebControls;
 using System.Data;
 using System.Data.SqlClient;
+using System.Text.RegularExpressions;
+
 /*
 界面权限定义
 员工信息查询：2
@@ -18,19 +20,21 @@ namespace sylzyb_employer_mgr
 {
     public partial class employer_mgr : System.Web.UI.Page
     {
-        public Check option_ck=new Check();
+        public Check option_ck = new Check();
         public static string sel_string = "";
-        public  static string option_sql= "";
+        public static string option_sql = "";
         public db db_opt = new db();
-        public SqlDataReader dr_select_row ;
+        public SqlDataReader dr_select_row;
         public DataSet ds = new DataSet();
-        static  string option_method = "";
+        static string option_method = "";
+        private int module_kind = 0;
+        private int item_kind = 2;
         protected void Page_Load(object sender, EventArgs e)
         {
             if (!IsPostBack)
             {
 
-                if (option_ck.moudle("员工信息管理") == false || System.Web.HttpContext.Current.Session["UserName"].ToString() == "" || System.Web.HttpContext.Current.Session["IDCard"] == null)
+                if (option_ck.Module("员工信息管理", module_kind) == false || System.Web.HttpContext.Current.Session["UserName"].ToString() == "" || System.Web.HttpContext.Current.Session["IDCard"] == null)
                 {
                     System.Web.HttpContext.Current.Session["RealName"] = "";
                     System.Web.HttpContext.Current.Session["IDCard"] = "";
@@ -40,7 +44,7 @@ namespace sylzyb_employer_mgr
                     System.Web.HttpContext.Current.Session["UserPower"] = "";
                     System.Web.HttpContext.Current.Session["ModulePower"] = "";
                     Page.ClientScript.RegisterStartupScript(this.GetType(), "", "<script language='javascript'>alert('您尚未登陆或登陆超时');location.href='Login.aspx';</script>");
-                   
+
                 }
                 else
                 {
@@ -48,14 +52,14 @@ namespace sylzyb_employer_mgr
                     sel_string = "select * from [dzsw].[dbo].[Syl_WorkerInfo] ";
                 }
             }
-            btn_emp_add.Visible = option_ck.item("员工-"+"添加员工信息");
-            btn_emp_edt.Visible = option_ck.item("员工-"+"修改员工信息");
-            btn_emp_del.Visible = option_ck.item("员工-"+"删除员工信息");
+            btn_emp_add.Visible = option_ck.item("添加员工信息", item_kind);
+            btn_emp_edt.Visible = option_ck.item("修改员工信息", item_kind);
+            btn_emp_del.Visible = option_ck.item("删除员工信息", item_kind);
 
 
             employer_detail.Visible = false;
             btn_ok.Visible = false;
-           btn_cancel.Visible = false;
+            btn_cancel.Visible = false;
 
             tbx_id.Enabled = false;
             tbx_WorkerName.Enabled = false;
@@ -66,12 +70,12 @@ namespace sylzyb_employer_mgr
             tbx_WagesFactor.Enabled = false;
             tbx_DutiesFactor.Enabled = false;
 
-            ds= db_opt.build_dataset(sel_string);
+            ds = db_opt.build_dataset(sel_string);
             GridView1.DataSource = ds;
             GridView1.DataBind();
-           
+
             login_user.Text = System.Web.HttpContext.Current.Session["RealName"].ToString();
-           
+
         }
 
         protected void btn_back_Click(object sender, EventArgs e)
@@ -79,10 +83,10 @@ namespace sylzyb_employer_mgr
             Response.Redirect("login.aspx");
         }
 
-      
+
         protected void GridView1_SelectedIndexChanged(object sender, EventArgs e)
         {
-           
+
             for (int i = 0; i < GridView1.Rows.Count; i++)
             {
                 GridView1.Rows[i].BackColor = System.Drawing.Color.White;
@@ -93,10 +97,10 @@ namespace sylzyb_employer_mgr
                 employer_detail.Visible = true;
                 GridView1.Rows[GridView1.SelectedIndex].BackColor = System.Drawing.Color.BlanchedAlmond;
                 string sel_rec = "";
-                sel_rec = "select * from  [dzsw].[dbo].[Syl_WorkerInfo]  where  ID='" + GridView1.Rows[GridView1.SelectedIndex].Cells[0].Text+"'";
-               dr_select_row=db_opt.datareader(sel_rec);
+                sel_rec = "select * from  [dzsw].[dbo].[Syl_WorkerInfo]  where  ID='" + GridView1.Rows[GridView1.SelectedIndex].Cells[0].Text + "'";
+                dr_select_row = db_opt.datareader(sel_rec);
                 while (dr_select_row.Read())
-                {                
+                {
                     tbx_id.Text = dr_select_row["ID"].ToString();
                     tbx_WorkerName.Text = dr_select_row["WorkerName"].ToString();
                     tbx_IDCard.Text = dr_select_row["IDCard"].ToString();
@@ -119,45 +123,81 @@ namespace sylzyb_employer_mgr
             String evt = Page.ClientScript.GetPostBackClientHyperlink(sender as GridView, "Select$" + e.Row.RowIndex.ToString());
             e.Row.Attributes.Add("onclick", evt);
         }
-
+ 
         protected void btn_ok_Click(object sender, EventArgs e)
         {
             btn_ok.Visible = false;
-            employer_detail.Visible = false;
+                     Boolean bool_isexist = db_opt.IsRecordExist("[dzsw].[dbo].[Syl_WorkerInfo]", "IDCard", tbx_IDCard.Text.Trim());
+            if (option_sql != "") option_sql = "";
             try
             {
+                Boolean bool_idcardFormat = Regex.IsMatch(tbx_IDCard.Text.Trim(), @"^\d{15}$") && tbx_IDCard.Text.Length == 15 || Regex.IsMatch(tbx_IDCard.Text.Trim(), @"^\d{17}[0-9Xx]$") && tbx_IDCard.Text.Length == 18;
 
-                if  (db_opt.IsRecordExist("[dzsw].[dbo].[Syl_WorkerInfo]", "WorkerName", tbx_WorkerName.Text ))
+                if (bool_idcardFormat == false)
                 {
-                    if (option_sql != "") option_sql = "";
-                    if (string.Compare(option_method, "insert") == 0)
-                        option_sql = "insert into [dzsw].[dbo].[Syl_WorkerInfo] (WorkerName,IDCard,GroupName,Job,Duties,WagesFactor,DutiesFactor)values('"
-                        + tbx_WorkerName.Text + "','" + tbx_IDCard.Text + "','" + tbx_GroupName.Text + "','" + tbx_Job.Text + "','" + tbx_Duties.Text + "','"
-                        + Convert.ToDecimal(tbx_WagesFactor.Text) + "','" + Convert.ToDecimal(tbx_DutiesFactor.Text) + "')";
-                    if (string.Compare(option_method, "delete") == 0)
-                        option_sql = "delete  from [dzsw].[dbo].[Syl_WorkerInfo] where  ID='" + tbx_id.Text + "'";
-                    if (string.Compare(option_method, "update") == 0)
-                        option_sql = "update  [dzsw].[dbo].[Syl_WorkerInfo] set WorkerName='"
-                             + tbx_WorkerName.Text.Trim() + "',IDCard='" + tbx_IDCard.Text.Trim() + "',GroupName='" + tbx_GroupName.Text.Trim()
-                             + "',Job='" + tbx_Job.Text.Trim() + "',Duties='" + tbx_Duties.Text.Trim() + "',WagesFactor='" + Convert.ToDecimal(tbx_WagesFactor.Text.Trim())
-                             + "',DutiesFactor='" + Convert.ToDecimal(tbx_DutiesFactor.Text.Trim()) + "' where id='" + tbx_id.Text.Trim() + "'";
-                    if (option_sql != "")
-                    {
-                        db_opt.execsql(option_sql);
-                        GridView1.DataSource = db_opt.build_dataset(sel_string);
-                        GridView1.DataBind();
-                        Page.ClientScript.RegisterStartupScript(this.GetType(), "message", "<script>alert('数据已经同步！');</script>");
-                    }
-                    else
-                        Page.ClientScript.RegisterStartupScript(this.GetType(), "message", "<script>alert('无效操作！');</script>");
+                    throw new Exception("身份证号: " + tbx_IDCard.Text.Trim());
                 }
-                else {
-                    Page.ClientScript.RegisterStartupScript(this.GetType(), "message", "<script>alert('该员工信息不存在于数据库！');</script>");
+                if (Regex.IsMatch(tbx_WagesFactor.Text, @"^\d{1,}$|^\d{1,}.\d{1,}$") == false)
+                {
+                    throw new Exception("工资系数: " + tbx_WagesFactor.Text);
                 }
+                if (Regex.IsMatch(tbx_DutiesFactor.Text, @"^\d{1,}$|^\d{1,}.\d{1,}$") == false)
+                {
+                    throw new Exception("管理奖系统:" + tbx_DutiesFactor.Text);
+                }
+
+
+
+                if (string.Compare(option_method, "insert") == 0 && !bool_isexist)
+                {
+                  
+                    option_sql = "insert into [dzsw].[dbo].[Syl_WorkerInfo] (WorkerName,IDCard,GroupName,Job,Duties,WagesFactor,DutiesFactor)values('"
+                    + tbx_WorkerName.Text + "','" + tbx_IDCard.Text.Trim() + "','"
+                    + tbx_GroupName.Text + "','" + tbx_Job.Text + "','" + tbx_Duties.Text + "','"
+                    + Convert.ToDecimal(tbx_DutiesFactor.Text.Trim()) + "','"
+                    + Convert.ToDecimal(tbx_DutiesFactor.Text.Trim()) + "')";
+                }
+                else
+                 if (string.Compare(option_method, "insert") == 0 && bool_isexist)
+                    throw new Exception("用户已经存在: " + tbx_IDCard.Text.Trim());
+
+                if (string.Compare(option_method, "delete") == 0 && bool_isexist)
+                    option_sql = "delete  from [dzsw].[dbo].[Syl_WorkerInfo] where  ID='" + tbx_id.Text + "'";
+                else
+                     if (string.Compare(option_method, "delete") == 0 && !bool_isexist)
+                    throw new Exception("用户不存在: " + tbx_IDCard.Text.Trim());
+
+                if (string.Compare(option_method, "update") == 0 && bool_isexist)
+                {
+                  
+                   option_sql = "update  [dzsw].[dbo].[Syl_WorkerInfo] set WorkerName='" + tbx_WorkerName.Text.Trim()
+                        + "',IDCard='" + tbx_IDCard.Text.Trim()
+                        + "',GroupName='" + tbx_GroupName.Text.Trim() + "',Job='" + tbx_Job.Text.Trim() + "',Duties='" + tbx_Duties.Text.Trim()
+                        + "',WagesFactor='" + Convert.ToDecimal(tbx_WagesFactor.Text.Trim())
+                        + "',DutiesFactor='" + Convert.ToDecimal(tbx_DutiesFactor.Text.Trim())
+                        + "' where id='" + tbx_id.Text.Trim() + "'";
+                }
+                else
+                     if (string.Compare(option_method, "update") == 0 && !bool_isexist)
+                    throw new Exception("用户不存在: " + tbx_IDCard.Text.Trim());
+
+                if (option_sql != "")
+                {
+                    db_opt.execsql(option_sql);
+                    GridView1.DataSource = db_opt.build_dataset(sel_string);
+                    GridView1.DataBind();
+                    Page.ClientScript.RegisterStartupScript(this.GetType(), "message", "<script>alert('数据已经同步！');</script>");
+                }
+                else
+                    Page.ClientScript.RegisterStartupScript(this.GetType(), "message", "<script>alert('无效操作！');</script>");
+
+
+                employer_detail.Visible = false;
             }
-            catch (Exception  opt_err)
+            catch (Exception opt_err)
             {
-                Page.ClientScript.RegisterStartupScript(this.GetType(), "message", "<script>alert('录入字段类型错误!"+opt_err.Message.ToString()+"');</script>");
+
+                Page.ClientScript.RegisterStartupScript(this.GetType(), "message", "<script>alert('录入字段类型或格式错误! " + opt_err.Message.ToString() + "');</script>");
             }
         }
 
@@ -180,7 +220,7 @@ namespace sylzyb_employer_mgr
             tbx_WagesFactor.Enabled = false;
             tbx_DutiesFactor.Enabled = false;
 
-           tbx_id.Text = "";
+            tbx_id.Text = "";
             tbx_WorkerName.Text = "";
             tbx_IDCard.Text = "";
             tbx_GroupName.Text = "";
@@ -189,8 +229,8 @@ namespace sylzyb_employer_mgr
             tbx_WagesFactor.Text = "";
             tbx_DutiesFactor.Text = "";
             // Response.Write("<script>alert('操作取消数据未同步！');javascript:history.go(-1);</script>");
-            Page.ClientScript.RegisterStartupScript(this.GetType(), "message", "<script>alert('操作取消数据未同步！"+e.ToString()+"');</script>");
-       
+            Page.ClientScript.RegisterStartupScript(this.GetType(), "message", "<script>alert('操作取消数据未同步！" + e.ToString() + "');</script>");
+
         }
 
         protected void btn_emp_add_Click(object sender, EventArgs e)
@@ -210,7 +250,7 @@ namespace sylzyb_employer_mgr
             tbx_WagesFactor.Enabled = true;
             tbx_DutiesFactor.Enabled = true;
 
-             tbx_id.Text = "";
+            tbx_id.Text = db_opt.max_id("id", "[dzsw].[dbo].[Syl_WorkerInfo]").ToString(); 
             tbx_WorkerName.Text = "";
             tbx_IDCard.Text = "";
             tbx_GroupName.Text = "";
@@ -255,7 +295,6 @@ namespace sylzyb_employer_mgr
             option_method = "update";
             btn_ok.Visible = true;
             btn_cancel.Visible = true;
-
             tbx_id.Enabled = false;
             tbx_WorkerName.Enabled = true;
             tbx_IDCard.Enabled = true;
@@ -265,16 +304,22 @@ namespace sylzyb_employer_mgr
             tbx_WagesFactor.Enabled = true;
             tbx_DutiesFactor.Enabled = true;
 
-            tbx_id.Text = GridView1.Rows[GridView1.SelectedIndex].Cells[0].Text== "&nbsp;" ? "" : GridView1.Rows[GridView1.SelectedIndex].Cells[0].Text;
-            tbx_WorkerName.Text = GridView1.Rows[GridView1.SelectedIndex].Cells[1].Text == "&nbsp;" ? "" : GridView1.Rows[GridView1.SelectedIndex].Cells[1].Text;
-            tbx_IDCard.Text = GridView1.Rows[GridView1.SelectedIndex].Cells[2].Text == "&nbsp;" ? "" : GridView1.Rows[GridView1.SelectedIndex].Cells[2].Text;
-            tbx_GroupName.Text = GridView1.Rows[GridView1.SelectedIndex].Cells[3].Text == "&nbsp;" ? "" : GridView1.Rows[GridView1.SelectedIndex].Cells[3].Text;
-            tbx_Job.Text = GridView1.Rows[GridView1.SelectedIndex].Cells[4].Text == "&nbsp;" ? "" : GridView1.Rows[GridView1.SelectedIndex].Cells[4].Text;
-            tbx_Duties.Text = GridView1.Rows[GridView1.SelectedIndex].Cells[5].Text == "&nbsp;" ? "" : GridView1.Rows[GridView1.SelectedIndex].Cells[5].Text;
-            tbx_WagesFactor.Text = GridView1.Rows[GridView1.SelectedIndex].Cells[6].Text == "&nbsp;" ? "" : GridView1.Rows[GridView1.SelectedIndex].Cells[6].Text;
-            tbx_DutiesFactor.Text = GridView1.Rows[GridView1.SelectedIndex].Cells[7].Text == "&nbsp;" ? "" : GridView1.Rows[GridView1.SelectedIndex].Cells[7].Text;
+
+            if (GridView1.SelectedIndex==0)
+                Page.ClientScript.RegisterStartupScript(this.GetType(), "message", "<script>alert('还没选择要编辑的员工！');</script>");
+            else
+            {
+                tbx_id.Text = GridView1.Rows[GridView1.SelectedIndex].Cells[0].Text == "&nbsp;" ? "" : GridView1.Rows[GridView1.SelectedIndex].Cells[0].Text;
+                tbx_WorkerName.Text = GridView1.Rows[GridView1.SelectedIndex].Cells[1].Text == "&nbsp;" ? "" : GridView1.Rows[GridView1.SelectedIndex].Cells[1].Text;
+                tbx_IDCard.Text = GridView1.Rows[GridView1.SelectedIndex].Cells[2].Text == "&nbsp;" ? "" : GridView1.Rows[GridView1.SelectedIndex].Cells[2].Text;
+                tbx_GroupName.Text = GridView1.Rows[GridView1.SelectedIndex].Cells[3].Text == "&nbsp;" ? "" : GridView1.Rows[GridView1.SelectedIndex].Cells[3].Text;
+                tbx_Job.Text = GridView1.Rows[GridView1.SelectedIndex].Cells[4].Text == "&nbsp;" ? "" : GridView1.Rows[GridView1.SelectedIndex].Cells[4].Text;
+                tbx_Duties.Text = GridView1.Rows[GridView1.SelectedIndex].Cells[5].Text == "&nbsp;" ? "" : GridView1.Rows[GridView1.SelectedIndex].Cells[5].Text;
+                tbx_WagesFactor.Text = GridView1.Rows[GridView1.SelectedIndex].Cells[6].Text == "&nbsp;" ? "" : GridView1.Rows[GridView1.SelectedIndex].Cells[6].Text;
+                tbx_DutiesFactor.Text = GridView1.Rows[GridView1.SelectedIndex].Cells[7].Text == "&nbsp;" ? "" : GridView1.Rows[GridView1.SelectedIndex].Cells[7].Text;
+            }
         }
 
-       
+
     }
 }
